finalVersion

+++ src/main/java/com/example/the_outer_market/model/Product.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public class Product {
@@ -13,26 +14,27 @@ public class Product {
     private Long id;
 
     private String name;
-    private int price;
+
+    private BigDecimal price;
+
     private int stock;
 
-    // Constructors
     public Product() {}
 
-    public Product(String name, int price, int stock) {
+    public Product(String name, double price, int stock) {
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
         this.stock = stock;
     }
 
     // Getters and Setters
     public Long getId() { return id; }
     public String getName() { return name; }
-    public int getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
     public int getStock() { return stock; }
 
     public void setId(Long id) { this.id = id; }
     public void setName(String name) { this.name = name; }
-    public void setPrice(int price) { this.price = price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
     public void setStock(int stock) { this.stock = stock; }
 }
